write unit test for GraphQLEntityViewSupport.createSetting
To facilitate easy test creation in the future, a helper class
is created to create types and mocks corresponding to the test
schema.

Also, graphql-java version is bumped to 10 since it contains
the classes/methods required for the test (namely SelectedField
and getFields) and the previosly used version 9.* is no longer
supported. Moreover, the current GraphQLEntityViewSupport code
doesn't work since 9.7 anyway.