[#1998] Upgrade to Quarkus 3.8 and switch to @ConfigMapping
At some point in the future, we will retire the legacy config classes so
let's migrate to the new @ConfigMapping infrastructure.

I was conservative and only updated to Quarkus 3.8.

Note that there might be more optimizations possible with having all
elements in a single map but I tried to be as less intrusive as possible
as I don't know the codebase.